Update default issue templates
This creates a default template for people to fill out when creating an issue on the project.   

Please let me know if you'd like to see something change.  This is really just a starting point, so let me know if you have any opinions about what should or should not be on this form.  
Thanks!